+++ src/config.ts
@@ -7,11 +7,11 @@ import { timeString } from "./helpers/time_helpers";
 import { type NonScoreWeightInfoDict, type Optional } from "./types";
 import {
     CacheKey,
+    FediverseCacheKey,
     LoadAction,
     LogAction,
     NonScoreWeightName,
     TagTootsCategory,
-    FEDIVERSE_CACHE_KEYS,
     type ApiCacheKey
 } from "./enums";
 
@@ -197,16 +197,16 @@ class Config implements ConfigType {
                 initialMaxRecords: Math.floor(MIN_RECORDS_FOR_FEATURE_SCORING / 2),  // Seems to be the biggest bottleneck
                 minutesUntilStale: 12 * MINUTES_IN_HOUR,
             },
-            [CacheKey.FEDIVERSE_POPULAR_SERVERS]: {
+            [FediverseCacheKey.FEDIVERSE_POPULAR_SERVERS]: {
                 minutesUntilStale: 5 * MINUTES_IN_DAY,
             },
-            [CacheKey.FEDIVERSE_TRENDING_LINKS]: {
+            [FediverseCacheKey.FEDIVERSE_TRENDING_LINKS]: {
                 minutesUntilStale: 4 * MINUTES_IN_HOUR,
             },
-            [CacheKey.FEDIVERSE_TRENDING_TAGS]: {
+            [FediverseCacheKey.FEDIVERSE_TRENDING_TAGS]: {
                 minutesUntilStale: 6 * MINUTES_IN_HOUR,
             },
-            [CacheKey.FEDIVERSE_TRENDING_TOOTS]: {
+            [FediverseCacheKey.FEDIVERSE_TRENDING_TOOTS]: {
                 minutesUntilStale: 4 * MINUTES_IN_HOUR,
             },
             [CacheKey.FOLLOWED_ACCOUNTS]: {
@@ -521,15 +521,17 @@ class Config implements ConfigType {
     };
 
     /**
-     * Computes the minimum value of minutesUntilStale for all FEDIVERSE_CACHE_KEYS.
+     * Computes the minimum value of minutesUntilStale for all FediverseCacheKey values.
      * Warns if any required keys are missing a value.
      * @returns {number} The minimum minutes until trending data is considered stale, or 60 if not all keys are configured.
      */
     minTrendingMinutesUntilStale(): number {
-        const trendStalenesses = FEDIVERSE_CACHE_KEYS.map(k => this.api.data[k]?.minutesUntilStale).filter(Boolean);
+        const trendStalenesses = Object.values(FediverseCacheKey)
+                                       .map(k => this.api.data[k]?.minutesUntilStale)
+                                       .filter(Boolean);
 
-        if (trendStalenesses.length != FEDIVERSE_CACHE_KEYS.length) {
-            console.warn(`${LOG_PREFIX} Not all FEDIVERSE_CACHE_KEYS have minutesUntilStale configured!`);
+        if (trendStalenesses.length != Object.values(FediverseCacheKey).length) {
+            console.warn(`${LOG_PREFIX} Not all FediverseCacheKey values have minutesUntilStale configured!`);
             return 60;
         } else {
             return Math.min(...trendStalenesses as number[]);